requirements.txt
requests
pillow
# Neccesary for resize-image-webp.py
webp